backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.16",
    "fastapi>=0.116.1",
    "google-genai>=1.29.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "mlxtend>=0.23.4",
    "networkx>=3.5",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "pyarrow>=21.0.0",
    "pypdf>=5.9.0",
    "redis>=6.4.0",
    "scikit-learn>=1.7.1",
    "tensorflow>=2.19.0",
    "uvicorn>=0.35.0",
    "xgboost>=3.0.3",
]
